pyproject.toml
[project]
name = "predibronx"
version = "0.1.0"
description = "AI-powered Polymarket prediction trading system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.52.0",
    "apscheduler>=3.11.2",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.6",
]
